adding single/single circular and double linked list code
