--- algorithm/src/com/codvary/sort/Printable.java
@@ -1,6 +1,0 @@
-package com.codvary.sort;
-
-public interface Printable {
-
-	public void print();
-}
