--- algorithm/src/com/codvary/sort/Sortable.java
@@ -1,7 +1,0 @@
-package com.codvary.sort;
-
-public interface Sortable extends Printable{
-
-	public void sort();
-	
-}
