+++ covarydsa/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>algorithm</name>
+	<name>covarydsa</name>
 	<comment></comment>
 	<projects>
 	</projects>
