+++ covarydsa/.settings/org.eclipse.core.resources.prefs
