+++ covarydsa/src/com/covary/dsa/linklist/DoubleLinkedList.java
@@ -1,6 +1,7 @@
-package com.codvary.linkedlist;
+package com.covary.dsa.linklist;
 
 public interface DoubleLinkedList extends SingleLinkedList {
 
 	public void displayReverse();
+
 }
