+++ covarydsa/src/com/covary/dsa/linklist/SingleCircularLinkList/SingleCircularLinkedListImpl.java
@@ -1,15 +1,14 @@
-package com.codvary.linkedlist.DCLList;
+package com.covary.dsa.linklist.SingleCircularLinkList;
 
-import com.codvary.linkedlist.DoubleLinkedList;
-import com.codvary.linkedlist.Node;
+import com.covary.dsa.linklist.Node;
+import com.covary.dsa.linklist.SingleLinkedList;
 
-public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
-
+public class SingleCircularLinkedListImpl implements SingleLinkedList {
 	private Node head;
 	private Node tail;
 	private int length;
 
-	public DoubleCircularLinkedListImpl() {
+	public SingleCircularLinkedListImpl() {
 		this.head = null;
 		this.tail = null;
 		this.length = 0;
@@ -18,13 +17,17 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 	@Override
 	public void display() {
 		if (this.head == null) {
-			System.out.print("List has no item!!!");
+			System.out.print("Empty List : Display");
 		} else {
 			Node iNode = this.head;
-			do {
-				System.out.print(iNode.getData() + " -> ");
+			while (iNode.getNext() != this.head) {
+				if (iNode.getNext() == this.head) {
+					System.out.print("[" + iNode.getData() + "]");
+				} else {
+					System.out.print("[" + iNode.getData() + "] -> ");
+				}
 				iNode = iNode.getNext();
-			} while (iNode != this.head);
+			}
 		}
 		System.out.println();
 	}
@@ -35,20 +38,16 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 		if (this.head == null) {
 			this.head = nNode;
 			this.tail = nNode;
-			this.tail.setNext(this.head);
+			nNode.setNext(this.head);
 		} else {
 			if (location == 0) {
 				nNode.setNext(this.head);
-				this.head.setPrev(nNode);
 				this.head = nNode;
-				this.head.setPrev(this.tail);
 				this.tail.setNext(this.head);
 			} else if (location >= this.length) {
 				this.tail.setNext(nNode);
-				nNode.setPrev(this.tail);
 				this.tail = nNode;
-				this.tail.setNext(this.head);
-				this.head.setPrev(this.tail);
+				nNode.setNext(this.head);
 			} else {
 				Node iNode = this.head;
 				for (int i = 0; i < location - 1; i++) {
@@ -56,9 +55,7 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 				}
 				Node nxNode = iNode.getNext();
 				nNode.setNext(nxNode);
-				nxNode.setPrev(nNode);
 				iNode.setNext(nNode);
-				nNode.setPrev(iNode);
 			}
 		}
 		this.length += 1;
@@ -67,7 +64,7 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 	@Override
 	public void removeNode(int location) {
 		if (this.head == null) {
-			System.out.println("List has no item!!!");
+			System.out.println("Empty List : Remove");
 		} else {
 			if (location == 0) {
 				if (this.head == this.tail) {
@@ -75,17 +72,19 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 					this.tail = null;
 				} else {
 					this.head = this.head.getNext();
-					this.head.setPrev(this.tail);
 					this.tail.setNext(this.head);
 				}
-			} else if (location >= this.length) {
+			} else if (location >= this.length - 1) {
 				if (this.head == this.tail) {
 					this.head = null;
 					this.tail = null;
 				} else {
-					this.tail = this.tail.getPrev();
-					this.tail.setNext(this.head);
-					this.head.setPrev(this.tail);
+					Node iNode = this.head;
+					while (iNode.getNext() != this.tail) {
+						iNode = iNode.getNext();
+					}
+					iNode.setNext(this.head);
+					this.tail = iNode;
 				}
 			} else {
 				Node iNode = this.head;
@@ -94,44 +93,27 @@ public class DoubleCircularLinkedListImpl implements DoubleLinkedList {
 				}
 				Node nxNode = iNode.getNext();
 				iNode.setNext(nxNode.getNext());
-				nxNode.getNext().setPrev(iNode);
 			}
 			this.length -= 1;
 		}
-
 	}
 
 	@Override
 	public boolean search(int data) {
 		if (this.head == null) {
-			System.out.println("List has no item!!!");
+			System.out.println("Empty List : Search");
 		} else {
 			Node iNode = this.head;
-			do {
+			while (iNode.getNext() != this.head) {
 				if (iNode.getData() == data) {
 					System.out.println("Found : [" + iNode.getData() + "]");
 					return true;
 				}
 				iNode = iNode.getNext();
-			} while (iNode != this.head);
-			System.out.println("Not Found : (" + data + ")");
+			}
+			System.out.println("Not Found : [" + data + "]");
 		}
 		return false;
 	}
 
-	@Override
-	public void displayReverse() {
-		if (this.head == null) {
-			System.out.print("List has no item!!!");
-		} else {
-			Node iNode = this.tail;
-			do {
-				System.out.print(iNode.getData() + " -> ");
-				iNode = iNode.getPrev();
-			}
-			while (iNode != this.tail);
-		}
-		System.out.println();
-
-	}
 }
