+++ covarydsa/src/com/covary/dsa/linklist/SingleLinkList/SingleLinkedListImpl.java
@@ -1,15 +1,15 @@
-package com.codvary.linkedlist.SLList;
+package com.covary.dsa.linklist.SingleLinkList;
 
-import com.codvary.linkedlist.Node;
-import com.codvary.linkedlist.SingleLinkedList;
+import com.covary.dsa.linklist.Node;
+import com.covary.dsa.linklist.SingleLinkedList;
 
-public class SingleLinkedListImple implements SingleLinkedList {
+public class SingleLinkedListImpl implements SingleLinkedList {
 
 	private Node head;
 	private Node tail;
 	private int length;
 
-	public SingleLinkedListImple() {
+	public SingleLinkedListImpl() {
 		this.head = null;
 		this.tail = null;
 		this.length = 0;
@@ -18,11 +18,15 @@ public class SingleLinkedListImple implements SingleLinkedList {
 	@Override
 	public void display() {
 		if (this.head == null) {
-			System.out.print("List has no item!!!");
+			System.out.print("Empty List : Display");
 		} else {
 			Node iNode = this.head;
 			while (iNode != null) {
-				System.out.print(iNode.getData()+" -> ");
+				if (iNode.getNext() == null) {
+					System.out.print("[" + iNode.getData() + "]");
+				} else {
+					System.out.print("[" + iNode.getData() + "] -> ");
+				}
 				iNode = iNode.getNext();
 			}
 		}
@@ -42,9 +46,9 @@ public class SingleLinkedListImple implements SingleLinkedList {
 			} else if (location >= this.length) {
 				this.tail.setNext(nNode);
 				this.tail = nNode;
-			}else{
+			} else {
 				Node iNode = this.head;
-				for(int i = 0 ; i < location; i++) {
+				for (int i = 0; i < location - 1; i++) {
 					iNode = iNode.getNext();
 				}
 				Node nxNode = iNode.getNext();
@@ -58,30 +62,30 @@ public class SingleLinkedListImple implements SingleLinkedList {
 	@Override
 	public void removeNode(int location) {
 		if (this.head == null) {
-			System.out.println("List has no item!!!");
+			System.out.println("Empty List : Remove");
 		} else {
 			if (location == 0) {
-				if(this.head == this.tail) {
+				if (this.head == this.tail) {
 					this.head = null;
 					this.tail = null;
-				}else {
+				} else {
 					this.head = this.head.getNext();
 				}
-			} else if (location >= this.length) {
-				if(this.head == this.tail) {
+			} else if (location >= this.length - 1) {
+				if (this.head == this.tail) {
 					this.head = null;
 					this.tail = null;
-				}else {
+				} else {
 					Node iNode = this.head;
-					while(iNode.getNext() != this.tail) {
+					while (iNode.getNext() != this.tail) {
 						iNode = iNode.getNext();
 					}
 					iNode.setNext(null);
 					this.tail = iNode;
 				}
-			}else{
+			} else {
 				Node iNode = this.head;
-				for(int i = 0; i < location - 1; i++) {
+				for (int i = 0; i < location - 1; i++) {
 					iNode = iNode.getNext();
 				}
 				Node nxNode = iNode.getNext();
@@ -93,18 +97,18 @@ public class SingleLinkedListImple implements SingleLinkedList {
 
 	@Override
 	public boolean search(int data) {
-		if(this.head == null) {
-			System.out.println("List has no item!!!");
-		}else {
+		if (this.head == null) {
+			System.out.println("Empty List : Search");
+		} else {
 			Node iNode = this.head;
-			while(iNode != null) {
-				if(iNode.getData() == data) {
-					System.out.println("Found : ["+iNode.getData()+"]");
+			while (iNode != null) {
+				if (iNode.getData() == data) {
+					System.out.println("Found : [" + iNode.getData() + "]");
 					return true;
 				}
 				iNode = iNode.getNext();
 			}
-			System.out.println("Not Found : ("+data+")");
+			System.out.println("Not Found : [" + data + "]");
 		}
 		return false;
 	}
