+++ covarydsa/src/com/covary/dsa/linklist/SingleLinkedList.java
@@ -1,13 +1,12 @@
-package com.codvary.linkedlist;
+package com.covary.dsa.linklist;
 
 public interface SingleLinkedList {
-	
-	public void display();
-	
-	public void addNode(int data, int location);
-	
-	public void removeNode(int location);
-	
-	public boolean search(int data);
 
+	public void display();
+
+	public void addNode(int data, int location);
+
+	public void removeNode(int location);
+
+	public boolean search(int data);
 }
